requirements.txt
django-filter~=22.1.0
djangorestframework-gis~=1.0.0
djangorestframework~=3.14.0
django~=3.2.3
psycopg2-binary~=2.9.0
django-leaflet~=0.29.0